Rename container to connector
